Added support for testing with kernel information
