+++ .cargo/config.toml
@@ -8,6 +8,6 @@ krun = "run --target x86_64-custom.json -Zbuild-std=core,alloc -Zbuild-std-featu
 kbuild-r = "build --target x86_64-custom.json -Zbuild-std=core,alloc --release -Zbuild-std-features=compiler-builtins-mem" #dev
 kimage-r = "run --target x86_64-custom.json -Zbuild-std=core,alloc --release -Zbuild-std-features=compiler-builtins-mem -- --no-run" #release
 krun-r = "run --target x86_64-custom.json -Zbuild-std=core,alloc --release -Zbuild-std-features=compiler-builtins-mem" #release
-ktest = "test --target x86_64-custom.json --bins -Zbuild-std=core,alloc -Zbuild-std-features=compiler-builtins-mem"
+ktest = "test --target x86_64-custom.json -Zbuild-std=core,alloc -Zbuild-std-features=compiler-builtins-mem"
 ktest_all = ["ktest"]
 kdoc = "doc --target x86_64-custom.json -Zbuild-std=core,alloc -Zbuild-std-features=compiler-builtins-mem"
