+++ kernel/src/debug.rs
@@ -1,6 +1,5 @@
 use bootloader::boot_info::{MemoryRegionKind, MemoryRegions};
-
-use test_framework::serial_println;
+use utils::serial_println;
 
 #[inline(always)]
 pub fn log(msg: &str) {
