+++ kernel/src/interrupts/cpu_handlers/general_protection_fault.rs
@@ -1,4 +1,4 @@
-use test_framework::serial_print;
+use utils::serial_print;
 use x86_64::structures::idt::InterruptStackFrame;
 
 /// Handles a general protection fault.
