+++ kernel/src/lib.rs
@@ -12,14 +12,8 @@
 extern crate alloc;
 
 use alloc::{boxed::Box, sync::Arc};
-use core::panic::PanicInfo;
 use lazy_static::lazy_static;
 use spin::Mutex;
-use test_framework::{
-    ansi_colors,
-    qemu_exit::{exit_qemu, QemuExitCode},
-    serial_println,
-};
 
 mod init;
 pub use init::{hlt_loop, init, register_driver, reload_drivers};
@@ -36,20 +30,4 @@ pub mod syscalls;
 
 lazy_static! {
     pub static ref LOGGER: Arc<Mutex<Option<Box<dyn Logger>>>> = Arc::from(Mutex::new(None));
-}
-
-/// Function called when a panic occurs
-pub fn panic_handler(info: &PanicInfo) -> ! {
-    serial_println!("{}\n", ansi_colors::Red("[PANIC]"));
-    serial_println!("Error: {}\n", info);
-    init::hlt_loop();
-}
-
-/// Function called when a panic while testing occurs
-pub fn test_panic_handler(info: &PanicInfo) -> ! {
-    serial_println!("{}\n", ansi_colors::Red("[PANIC]"));
-    serial_println!("Error: {}\n", info);
-    exit_qemu(QemuExitCode::Failed);
-
-    init::hlt_loop();
 }
