+++ test_framework/Cargo.toml
@@ -7,6 +7,5 @@ edition = { workspace=true }
 
 [dependencies]
 x86_64 = { workspace=true }
-spin = { workspace=true }
-uart_16550 = { workspace=true }
-lazy_static = { workspace=true }
+utils = { workspace=true }
+kernel = { workspace=true }
