+++ test_framework/src/lib.rs
@@ -3,6 +3,5 @@
 
 pub mod ansi_colors;
 pub mod qemu_exit;
-pub mod serial;
 pub mod test_runner;
 pub mod testable;
