+++ test_framework/src/testable.rs
@@ -1,19 +1,21 @@
+use kernel::structures::kernel_information::KernelInformation;
+use utils::{serial_print, serial_println};
+
 use crate::ansi_colors::Green;
-use crate::{serial_print, serial_println};
 
 /// Self documenting test runner trait
 pub trait Testable {
-    fn run(&self);
+    fn run(&self, kernel_information: KernelInformation);
 }
 
 impl<T> Testable for T
 where
-    T: Fn(),
+    T: Fn(KernelInformation),
 {
     /// Runs the test and prints the result
-    fn run(&self) {
+    fn run(&self, kernel_information: KernelInformation) {
         serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
+        self(kernel_information);
         serial_println!("{}", Green("[ok]"));
     }
 }
