+++ utils/Cargo.toml
@@ -7,3 +7,6 @@ edition = { workspace=true }
 
 [dependencies]
 x86_64 = { workspace=true }
+spin = { workspace=true }
+uart_16550 = { workspace=true }
+lazy_static = { workspace=true }
